Migrate to wikitable table class for table output
This extension previously used a custom table design that did not align with the
MediaWiki CSS style for tables. This change almost entirely rewrites the table
implementation to use the wikitable class, as well as enhances the readability of
the table by using proper table headings and moves the total number edits row to
the bottom of the table. While the existing TODO for inline styles was not entirely
followed, using the "text-align:center;" CSS property helps, in my mind, improve the
readability of the table, to the point where I feel that the TODO can still be removed.

Translation strings have been added to reflect the additional strings, and I have
additionally added myself to the authors array for both the extension information
file and the English translation file. Another change was made as well where I added
an output message before the OOUI form displaying a short description as to what the
extension does, thereby making it easier for users who stumble across this extension
to understand what exactly it does.

Change-Id: I0d4465b14f0de6425dafdcd52e16e14bb04a0541